Object-orriented
We added a table to the code and a new file that holds our class OnlineStore. We added code to connect and close hour database while using objects.Then we made code to foreach our error message to display all messages. We also added private variables to the class file.